install typescript and config
