+++ tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "include": [
+      "src/*"
+  ],
+  "compilerOptions": {
+      "target": "es5",
+      "jsx": "react",
+      "allowSyntheticDefaultImports": true
+  }
+}
